maintenance

+++ bin/pastmonth.sh
@@ -7,18 +7,32 @@ HERE=$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)
 pushd "${HERE}" >/dev/null || exit 1
     # shellcheck disable=SC1091
     source ./constants.sh
-    ./trend.py --days 0
 
     CURRENT_EPOCH=$(date +'%s')
     # do some maintenance
     # shellcheck disable=SC2154
-    echo -n "${db_full_path} integrity check: "
-    sqlite3 "${db_full_path}" "PRAGMA integrity_check;"
+    echo "${db_full_path} re-indexing... "
     sqlite3 "${db_full_path}" "REINDEX;"
 
-    # Keep upto 10 years of data
-    PURGE_EPOCH=$(echo "${CURRENT_EPOCH} - (10 * 366 * 24 * 3600)" | bc)
-    sqlite3 "${db_full_path}" \
-            "DELETE FROM kamstrup WHERE sample_epoch < ${PURGE_EPOCH};"
+    echo -n "${db_full_path} integrity check:   "
+    chk_result=$(sqlite3 "${db_full_path}" "PRAGMA integrity_check;")
+    echo " ${chk_result}"
+    if [ "${chk_result}" == "ok" ]; then
+        echo "${db_full_path} copying... "
+        # shellcheck disable=SC2154
+        cp "${db_full_path}" "${database_path}/backup/"
+
+        # Keep upto 10 years of data
+        echo "${db_full_path} vacuuming... "
+        PURGE_EPOCH=$(echo "${CURRENT_EPOCH} - (3660 * 24 * 3600)" |bc)
+        sqlite3 "${db_full_path}" \
+                "DELETE FROM kamstrup WHERE sample_epoch < ${PURGE_EPOCH};"
+        sqlite3 "${db_full_path}" \
+                "DELETE FROM production WHERE sample_epoch < ${PURGE_EPOCH};"
+        sqlite3 "${db_full_path}" \
+                "DELETE FROM zappi WHERE sample_epoch < ${PURGE_EPOCH};"
+    fi
+
+    ./trend.py --days 0
 
 popd >/dev/null || exit
